drop testpart/getpart tool, simplify building standalone
