+++ tests/server/.gitignore
@@ -9,5 +9,4 @@ rtspd
 sockfilt
 socksd
 sws
-testpart
 tftpd
